Revert the last commit

+++ kornia/filters/filter.py
@@ -123,7 +123,7 @@ def filter2d(
 
     # convolve the tensor with the kernel.
     # NOTE: type(...) to fix getting `torch.bfloat16` type.
-    output = F.conv2d(input, tmp_kernel, groups=tmp_kernel.size(0), padding=0, stride=1).type(input.dtype)
+    output = F.conv2d(input, tmp_kernel, groups=tmp_kernel.size(0), padding=0, stride=1)
 
     if padding == 'same':
         out = output.view(b, c, h, w)
